requirements.txt
google-adk[a2a]>=0.1.0
google-cloud-aiplatform>=1.50.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
opentelemetry-instrumentation-google-genai>=0.1.0